# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeHub - Universal Coding Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile and connected platform handles
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    post:
      operationId: createUserProfile
      tags: [user]
      summary: Create user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── PLATFORMS ───────────────────────────────────────────────────────────────
...
  /platforms/{platformId}/connect:
    post:
      operationId: connectPlatform
      tags: [platforms]
      summary: Connect a platform with username or profile URL
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformConnection"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStatus"
        "400":
          description: Invalid credentials
  /platforms/{platformId}/disconnect:
    delete:
      operationId: disconnectPlatform
      tags: [platforms]
      summary: Disconnect a platform
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Disconnected
  /platforms/{platformId}/sync:
    post:
      operationId: syncPlatform
      tags: [platforms]
      summary: Trigger a manual sync for a platform
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
  /platforms/{platformId}/stats:
    get:
      operationId: getPlatformStats
      tags: [platforms]
      summary: Get detailed stats for a specific platform
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformDetailStats"
        "404":
          description: Platform not connected
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, displayName, onboardingComplete, platforms, createdAt]
      properties:
        id:
          type: integer
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformHandle"
        createdAt:
          type: string
          format: date-time
        lastSyncAt:
          type: ["string", "null"]
          format: date-time
    UserProfileInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
        avatarUrl:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformHandle"
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        avatarUrl:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformHandle"
    PlatformHandle:
      type: object
      required: [platformId, handle]
      properties:
        platformId:
          type: string
        handle:
          type: string

    # ─── PLATFORMS ────────────────────────────────────────────────────────────
    PlatformStatus:
      type: object
      required: [id, name, slug, connected, color, logoUrl]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        connected:
          type: boolean
        handle:
          type: ["string", "null"]
        color:
          type: string
        logoUrl:
          type: string
        lastSyncAt:
          type: ["string", "null"]
          format: date-time
        syncStatus:
          type: ["string", "null"]
          enum: [idle, syncing, error, null]
        totalSolved:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
    PlatformConnection:
      type: object
      required: [handle]
      properties:
        handle:
          type: string
          minLength: 1
    SyncResult:
      type: object
      required: [platformId, status, newSubmissions]
      properties:
        platformId:
          type: string
        status:
          type: string
          enum: [success, partial, error]
        newSubmissions:
          type: integer
        errorMessage:
          type: ["string", "null"]
        syncedAt:
          type: string
          format: date-time
...
    PlatformDetailStats:
      type: object
      required: [platformId, name, totalSolved, easySolved, mediumSolved, hardSolved]
      properties:
        platformId:
          type: string
        name:
          type: string
        handle:
          type: string
        totalSolved:
          type: integer
        easySolved:
          type: integer
        mediumSolved:
          type: integer
        hardSolved:
          type: integer
        totalSubmissions:
          type: integer
        acceptanceRate:
          type: ["number", "null"]
        ranking:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        streak:
          type: ["integer", "null"]
        maxStreak:
          type: ["integer", "null"]
        contestsParticipated:
          type: ["integer", "null"]
        activeDays:
          type: integer
        topicsBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/TopicCount"
        languagesUsed:
          type: array
          items:
            $ref: "#/components/schemas/LanguageCount"
        recentSubmissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        submissionGraph:
          type: array
          items:
            $ref: "#/components/schemas/DayCount"

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
    Submission:
      type: object
      required: [id, platformId, platformName, problemName, solvedAt]
      properties:
        id:
          type: integer
        platformId:
          type: string
        platformName:
          type: string
        problemName:
          type: string
        problemUrl:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [easy, medium, hard, null]
        topic:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        status:
          type: string
          enum: [accepted, wrong_answer, time_limit, runtime_error, compilation_error, unknown]
        isContest:
          type: boolean
        solvedAt:
          type: string
          format: date-time
        timeTakenMs:
          type: ["integer", "null"]
        companyTags:
          type: array
          items:
            type: string
...
    TopicCount:
      type: object
      required: [topic, count]
      properties:
        topic:
          type: string
        count:
          type: integer
    LanguageCount:
      type: object
      required: [language, count]
      properties:
        language:
          type: string
        count:
          type: integer
    DayCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
...
    Badge:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        iconUrl:
          type: ["string", "null"]
...